"Bump" Angular to 1.4 Release Candidate
It’s not really a bump since we have to specifically specify that we
need to use the master branch of the Angular repository.

We might have to “bower update” our installations more often, but the
features in ng1.4 are not to be missed out on.

I also removed the dev dependencies as we dont write any tests.

geiles merges geiles