I know, a confusing PR with outcommented functions and shit.
But it does take care of some stuff though.

First of all, there is now a full flow (except a bug, ill explain later)

So, the new flow goes like this.

on url: '/'
The app opens the overview.
When clicking a book, fire up a to cool for school animation and when
the animation is done, it redirects to single view ('/documents/:id').
On single view there is now a delete button, where if clicked, will
destroy the document and redirect to overview (url: '/').

Then also the user can delete single entry's on the overview page
(wasn't possible before as a click event listener to run the animation
took the click and opened up the document, as the same time deleting it,
but still having it in vm.current... i know, mindfuck).
BUT, the list isn't updatede until reload og stateChange (i know there's
something there i don't understand which i'm looking forward to
understand).

Though if all documents are deleted, it won't create a new one anymore
(i thought it was the coolest feature), but when the overview is not in
the sidebar, it felt extremely confusing!

And a create new link is now in the overview page as well :)

Pull master and enjoy :)