In this commit i have disabled foundations styling for buttons, and instead used their plugin to generate som button styling.
There are three sizes:
In this commit i have disabled foundations styling for buttons, and
instead used their plugin to generate som button styling.

There are three sizes:
.button--small
.button (default)
.button--large

Also there are 2 styles
.button (default, has a grey background)
.button.button--clear (which has no background and a grey color)

Issues
There are some issues related to the grunticon plugin which i haven't
yet been able to solve, but since it's already in the generator
@dennishn you might have some ideas about how to solve my issue.

I would like to embed the svg code to the dom, and trough that be able
to style the svg with css (right now the color can't be changed), it
works fine if i just do it in index.html, but not with dynamicly
rendered DOM, see issue
https://github.com/filamentgroup/grunticon/issues/258
Where a guy refere to som helper functions, which i don't underdstand
how to use. so before i dig deeper into it, maybe you know something i
don't @dennishn.