Removing $timeouts
Biggest factor is that instead of using a scope registered dom object,
we use the raw template object generated by the $compile method.

What $timeout gave you was a digest cycle. There’s no need to wait for
that when we _know_ that the element is in the dom.

Also fixed the issue where the dropdown was not removed from the dom.